@@ -12,21 +12,20 @@
 #   specific language governing permissions and limitations under the License.
 # ---  END LICENSE_HEADER BLOCK  ---
 
-require 'avalon/controller/controller_behavior'
 require 'avalon/intercom'
 
 class MediaObjectsController < ApplicationController
   include Rails::Pagination
-  include Avalon::Workflow::WorkflowControllerBehavior
-  include Avalon::Controller::ControllerBehavior
+  include WorkflowControllerBehavior
+  include DownloadBehavior
   include ConditionalPartials
   include NoidValidator
   include SecurityHelper
 
   before_action :authenticate_user!, except: [:show, :set_session_quality, :show_stream_details, :manifest]
-  before_action :load_resource, except: [:create, :destroy, :update_status, :set_session_quality, :tree, :deliver_content, :confirm_remove, :show_stream_details, :add_to_playlist, :intercom_collections, :manifest, :move_preview, :edit, :update, :json_update]
-  load_and_authorize_resource except: [:create, :destroy, :update_status, :set_session_quality, :tree, :deliver_content, :confirm_remove, :show_stream_details, :add_to_playlist, :intercom_collections, :manifest, :move_preview, :show_progress]
-  authorize_resource only: [:create]
+  before_action :load_resource, except: [:create, :destroy, :update_status, :set_session_quality, :tree, :deliver_content, :confirm_remove, :show_stream_details, :add_to_playlist, :intercom_collections, :manifest, :move_preview, :update, :json_update]
+  load_and_authorize_resource except: [:create, :destroy, :update_status, :set_session_quality, :tree, :deliver_content, :confirm_remove, :show_stream_details, :add_to_playlist, :intercom_collections, :manifest, :move_preview, :show_progress, :edit]
+  authorize_resource only: [:create, :edit]
 
   before_action :inject_workflow_steps, only: [:edit, :update], unless: proc { request.format.json? }
   before_action :load_player_context, only: [:show]
@@ -109,22 +108,24 @@ class MediaObjectsController < ApplicationController
 
   # POST /media_objects/avalon:1/add_to_playlist
   def add_to_playlist
-    @media_object = SpeedyAF::Proxy::MediaObject.find(params[:id])
+    add_to_playlist_params = params.require(:post).permit(:masterfile_id, :playlist_id, :playlistitem_scope)
+    @media_object = add_to_playlist_params.present? ? SpeedyAF::Proxy::MediaObject.find(params[:id]) : SpeedyAF::Proxy::MediaObject.find(params[:id], load_reflections: true)
     authorize! :read, @media_object
     masterfile_id = params[:post][:masterfile_id]
     playlist_id = params[:post][:playlist_id]
     playlist = Playlist.find(playlist_id)
+    new_items = []
     if current_ability.cannot? :update, playlist
       render json: {message: "<p>You are not authorized to update this playlist.</p>", status: 403}, status: 403 and return
     end
     playlistitem_scope = params[:post][:playlistitem_scope] #'section', 'structure'
     # If a single masterfile_id wasn't in the request, then create playlist_items for all masterfiles
-    masterfile_ids = masterfile_id.present? ? [masterfile_id] : @media_object.section_ids
-    masterfile_ids.each do |mf_id|
-      mf = SpeedyAF::Proxy::MasterFile.find(mf_id)
-      if playlistitem_scope=='structure' && mf.has_structuralMetadata? && mf.structuralMetadata.xpath('//Span').present?
+    masterfiles = masterfile_id.present? ? [SpeedyAF::Proxy::MasterFile.find(masterfile_id)] : @media_object.sections
+    masterfiles.each do |mf|
+      sf = mf.has_structuralMetadata? ? mf.structuralMetadata : nil
+      if playlistitem_scope=='structure' && sf.present? && sf.xpath('//Span').present?
         #create individual items for spans within structure
-        mf.structuralMetadata.xpath('//Span').each do |s|
+        sf.xpath('//Span').each do |s|
           labels = [mf.embed_title]
           labels += s.xpath('ancestor::Div[\'label\']').collect{|a|a.attribute('label').value.strip}
           labels << s.attribute('label')
@@ -134,16 +135,16 @@ class MediaObjectsController < ApplicationController
           start_time = time_str_to_milliseconds(start_time.value) if start_time.present?
           end_time = time_str_to_milliseconds(end_time.value) if end_time.present?
           clip = AvalonClip.new(title: label, master_file: mf, start_time: start_time, end_time: end_time)
-          new_item = PlaylistItem.new(clip: clip, playlist: playlist)
-          playlist.items += [new_item]
+          new_items += [PlaylistItem.new(clip: clip, playlist: playlist)]
         end
       else
         #create a single item for the entire masterfile
         item_title = @media_object.section_ids.count > 1 ? mf.embed_title : @media_object.title
         clip = AvalonClip.new(title: item_title, master_file: mf)
-        playlist.items += [PlaylistItem.new(clip: clip, playlist: playlist)]
+        new_items += [PlaylistItem.new(clip: clip, playlist: playlist)]
       end
     end
+    playlist.items += new_items
     link = view_context.link_to('View Playlist', playlist_path(playlist), class: "btn btn-primary btn-sm")
     render json: {message: "<p>Playlist items created successfully.</p> #{link}", status: 200}
   end
@@ -240,7 +241,8 @@ class MediaObjectsController < ApplicationController
           break
         end
         if file_spec[:files].present?
-          if master_file.update_derivatives(file_spec[:files], false)
+          master_file.update_derivatives(file_spec[:files], false)
+          if master_file.save
             master_file.update_stills_from_offset!
             WaveformJob.perform_later(master_file.id)
           else
