@@ -37,7 +37,8 @@ const Ramp = ({
   playlist,
   cdl,
   has_files,
-  has_transcripts
+  has_transcripts,
+  accessibility_text = '',
 }) => {
   const [manifestUrl, setManifestUrl] = React.useState('');
   const [startCanvasId, setStartCanvasId] = React.useState();
@@ -66,21 +67,25 @@ const Ramp = ({
     setManifestUrl(url);
   }, []);
 
+  const a11yWithOnlyShare = React.useMemo(() => {
+    return accessibility_text && !(timeline.canCreate && playlist.canCreate);
+  }, [accessibility_text, timeline.canCreate, playlist.canCreate]);
+
   return (
     <IIIFPlayer manifestUrl={manifestUrl}
       customErrorMessage='This page encountered an error. Please refresh or contact an administrator.'
       startCanvasId={startCanvasId}
       startCanvasTime={startCanvasTime}>
       <Row className="ramp--all-components ramp--itemview">
-        <Col sm={12} md={9} lg={8}>
+        <Col sm={12} md={12} xl={8}>
           {(cdl.enabled && !cdl.can_stream)
             ? (<React.Fragment>
-                <div dangerouslySetInnerHTML={{ __html: cdl.embed }} />
-                <div className="ramp--rails-title">
-                  {<div className="object-title" dangerouslySetInnerHTML={{ __html: title.content }} />}
-                </div>
-              </React.Fragment>
-              )
+              <div dangerouslySetInnerHTML={{ __html: cdl.embed }} />
+              <div className="ramp--rails-title">
+                {<div className="object-title" dangerouslySetInnerHTML={{ __html: title.content }} />}
+              </div>
+            </React.Fragment>
+            )
             : (<React.Fragment>
               {sections_count > 0 &&
                 <React.Fragment>
@@ -88,8 +93,8 @@ const Ramp = ({
                   <div className="ramp--rails-title">
                     {<div className="object-title" dangerouslySetInnerHTML={{ __html: title.content }} />}
                   </div>
-                  <div className="ramp--rails-content">
-                    <Col className="ramp-button-group-1">
+                  <div className={`ramp--rails-content ${a11yWithOnlyShare ? 'only-share' : ''}`}>
+                    <Col className="ramp-button-group-1" sm={accessibility_text ? 6 : 12} xs={a11yWithOnlyShare ? 4 : 12}>
                       {timeline.canCreate &&
                         <button
                           id="timelineBtn"
@@ -101,7 +106,6 @@ const Ramp = ({
                           aria-controls="timelineModal"
                           disabled={true}
                           data-testid="media-object-create-timeline-btn"
-                        
                         >
                           Create Timeline
                         </button>
@@ -115,6 +119,7 @@ const Ramp = ({
                           aria-expanded="false"
                           aria-controls="shareResourcePanel"
                           id="shareBtn"
+                          data-testid="media-object-share-btn"
                         >
                           <i className="fa fa-share-alt"></i>
                           Share
@@ -140,6 +145,11 @@ const Ramp = ({
                         </button>
                       }
                     </Col>
+                    {accessibility_text &&
+                      <Col className='accessibility-request text-right' sm={6} xs={a11yWithOnlyShare ? 8 : 12}>
+                        <span dangerouslySetInnerHTML={{ __html: accessibility_text }} />
+                      </Col>
+                    }
                   </div>
                   <Row className="mx-0">
                     <Col>
@@ -163,7 +173,7 @@ const Ramp = ({
             )
           }
         </Col>
-        <Col sm={12} md={3} lg={4} className="ramp--tabs-panel">
+        <Col sm={12} md={12} xl={4} className="ramp--tabs-panel">
           {cdl.enabled && <div dangerouslySetInnerHTML={{ __html: cdl.destroy }} />}
           <Tabs>
             <Tab eventKey="details" title="Details" >
