@@ -58,6 +58,9 @@ class MasterFile < ActiveFedora::Base
   property :title, predicate: ::RDF::Vocab::EBUCore.title, multiple: false do |index|
     index.as :stored_searchable
   end
+  property :original_filename, predicate: Avalon::RDFVocab::MasterFile.originalFileName, multiple: false do |index|
+    index.as :stored_sortable
+  end
   property :file_location, predicate: Avalon::RDFVocab::EBUCore.locator, multiple: false do |index|
     index.as :stored_sortable
   end
@@ -218,6 +221,7 @@ class MasterFile < ActiveFedora::Base
     else #Batch
       saveOriginal(file, File.basename(file.path), dropbox_dir)
     end
+    self.original_filename = File.basename(file_name) unless (file.is_a?(Hash) || file_name.nil?)
 
     @auth_header = auth_header
     reloadTechnicalMetadata!
@@ -314,11 +318,11 @@ class MasterFile < ActiveFedora::Base
     # We can get the proper aspect ratio from the transcoded files, so we set the master file off the 
     # encode output.
     if is_video?
-      high_output = Array(encode.output).select { |out| out.label.include?("high") }.first
+      high_output = Array(encode.output).select { |out| out.label&.include?("high") }.first
       self.display_aspect_ratio = (high_output.width.to_f / high_output.height.to_f).to_s
     end
 
-    outputs = Array(encode.output).collect do |output|
+    outputs = Array(encode.output).reject { |output| output.format == "vtt" }.collect do |output|
       {
         id: output.id,
         label: output.label,
@@ -335,6 +339,13 @@ class MasterFile < ActiveFedora::Base
       }
     end
     update_derivatives(outputs)
+
+    supplemental_file_outputs = Array(encode.output).select { |out| out.format == 'vtt' }
+    supplemental_file_ids = supplemental_file_outputs.collect { |sf| sf.id }.compact
+    add_supplemental_files(supplemental_file_ids) if supplemental_file_ids.present?
+
+    save
+
     run_hook :after_transcoding
   end
 
@@ -348,8 +359,10 @@ class MasterFile < ActiveFedora::Base
         existing.delete
       end
     end
+  end
 
-    save
+  def add_supplemental_files(ids)
+    self.supplemental_files += ids.collect { |id| GlobalID::Locator.locate(id) }
   end
 
   alias_method :'_poster_offset', :'poster_offset'
@@ -724,6 +737,7 @@ class MasterFile < ActiveFedora::Base
       next unless usable_files.has_key?(quality)
       self.file_location = File.realpath(usable_files[quality])
       self.file_size = usable_files[quality].size.to_s
+      self.original_filename = File.basename(File.realpath(usable_files[quality]))
       break
     end
   ensure
