@@ -18,7 +18,7 @@ class MediaObject < ActiveFedora::Base
   include VirtualGroups
   include ActiveFedora::Associations
   include MediaObjectMods
-  include Avalon::Workflow::WorkflowModelMixin
+  include WorkflowModelMixin
   include Permalink
   include Identifier
   include MigrationTarget
@@ -99,7 +99,7 @@ class MediaObject < ActiveFedora::Base
   end
 
   def validate_language
-    Array(language).each{|i|errors.add(:language, "Language not recognized (#{i[:code]})") unless LanguageTerm::map[i[:code]] }
+    Array(language).each{|i|errors.add(:language, "Language not recognized (#{i[:code]})") unless LanguageTerm::Iso6392.map[i[:code]] }
   end
 
   def validate_related_items
@@ -168,8 +168,14 @@ class MediaObject < ActiveFedora::Base
     @sections = mfs
   end
 
-  def sections
-    @sections ||= MasterFile.find(self.section_ids)
+  def sections(safe_load: false)
+    if safe_load
+      # Trick ActiveFedora to use solr so it doesn't raise an exception when a section has already been deleted
+      # In this case return all sections that were found in both solr and fedora
+      @sections = MasterFile.where("has_model_ssim:MasterFile").find(self.section_ids).compact
+    else
+      @sections ||= MasterFile.find(self.section_ids)
+    end
   end
 
   def section_ids
@@ -187,11 +193,15 @@ class MediaObject < ActiveFedora::Base
   end
 
   def destroy
+    # Ignore sections that have already been deleted (or don't appear in solr)
+    existing_sections = self.sections(safe_load: true)
+
     # attempt to stop the matterhorn processing job
-    self.sections.each(&:stop_processing!)
+    existing_sections.each(&:stop_processing!)
     # avoid calling destroy on each section since it calls save on parent media object
-    self.sections.each(&:delete)
+    existing_sections.each(&:delete)
     Bookmark.where(document_id: self.id).destroy_all
+    Checkout.where(media_object_id: self.id).destroy_all
     super
   end
 
@@ -343,6 +353,7 @@ class MediaObject < ActiveFedora::Base
       #Add all searchable fields to the all_text_timv field
       all_text_values = []
       all_text_values << solr_doc["title_tesi"]
+      all_text_values << solr_doc["alternative_title_ssim"]
       all_text_values << solr_doc["creator_ssim"]
       all_text_values << solr_doc["contributor_ssim"]
       all_text_values << solr_doc["unit_ssim"]
@@ -496,4 +507,9 @@ class MediaObject < ActiveFedora::Base
       # TODO: Optimize this into a single solr query?
       section_ids.select { |m| SpeedyAF::Proxy::MasterFile.find(m).supplemental_files(tag: tag).present? }
     end
+
+    def sections_with_rendering_files?(tags)
+      tags.any? { |t| sections_with_files(tag: t).present? }
+    end
+
 end
