@@ -32,23 +32,45 @@ class WatchedEncode < ActiveEncode::Base
     if Settings.encoding.derivative_bucket &&
        (Settings.encoding.engine_adapter.to_sym == :ffmpeg || is_a?(PassThroughEncode))
       bucket = Aws::S3::Bucket.new(name: Settings.encoding.derivative_bucket)
+
       encode.output.collect! do |output|
-        file = FileLocator.new output.url
-        key = file.location.sub(/\/(.*?)\//, "")
-        obj = bucket.object key
-
-        if File.exist? file.location
-          obj.upload_file file.location
-          File.delete file.location
+        if output.format == "vtt"
+          new_file = SupplementalFile.new(tags: ['caption'], parent_id: record.master_file_id)
+          new_file.label = output.label.presence
+          if output.language.present?
+            begin
+              new_file.language = LanguageTerm.find(output.language).code
+            rescue LanguageTerm::LookupError
+              new_file.language = nil
+            end
+          end
+          new_file.attach_file(FileLocator.new(output.url).location, io: true)
+          new_file.save
+          output.url = if Settings.active_storage.bucket.present?
+                      "s3://#{Settings.active_storage.bucket}/#{new_file.file.blob.key}"
+                    else
+                      new_file.file.blob.url
+                    end
+          output.id = new_file.to_global_id.to_s
+          output
         else
-          # Calls to Addressable::URI.escape here are to counter the unescaping that happens in FileLocator
-          # This is needed because files uploaded to minio (and probably AWS) escape spaces (%20) instead of keeping spaces
-          obj.upload_file Addressable::URI.escape(file.location)
-          File.delete Addressable::URI.escape(file.location)
-        end
+          file = FileLocator.new output.url
+          key = file.location.sub(/\/(.*?)\//, "")
+          obj = bucket.object key
 
-        output.url = "s3://#{obj.bucket.name}/#{obj.key}"
-        output
+          if File.exist? file.location
+            obj.upload_file file.location
+            File.delete file.location
+          else
+            # Calls to Addressable::URI.escape here are to counter the unescaping that happens in FileLocator
+            # This is needed because files uploaded to minio (and probably AWS) escape spaces (%20) instead of keeping spaces
+            obj.upload_file Addressable::URI.escape(file.location)
+            File.delete Addressable::URI.escape(file.location)
+          end
+
+          output.url = "s3://#{obj.bucket.name}/#{obj.key}"
+          output
+        end
       end
 
       # Save translated output urls
