@@ -1,6 +1,7 @@
 require 'blacklight'
+
 Rails.application.config.to_prepare do
-  module Blacklight::UrlHelperBehavior
+  Blacklight::SearchState.class_eval do
     def url_for_document doc, options = {}
       SpeedyAF::Base.for(doc.to_h.with_indifferent_access)
     end
