@@ -3,10 +3,10 @@ config = Rails.application.config
 redis_host = Settings.redis.host
 redis_port = Settings.redis.port || 6379
 redis_db = Settings.redis.db || 0
-config.cache_store = :redis_store, {
-  host: redis_host,
-  port: redis_port,
-  db: redis_db,
+
+redis_url = Settings.redis.url || "redis://#{redis_host}:#{redis_port}/#{redis_db}"
+config.cache_store = :redis_cache_store, {
+  url: redis_url,
   namespace: 'avalon'
 }
 
