@@ -1,4 +1,5 @@
-redis_conn = { url: "redis://#{Settings.redis.host}:#{Settings.redis.port}/#{Settings.redis.db}" }
+redis_url = Settings.redis.url || "redis://#{Settings.redis.host}:#{Settings.redis.port}/#{Settings.redis.db}"
+redis_conn = { url: redis_url }
 Sidekiq.configure_server do |s|
   s.redis = redis_conn
 end
@@ -21,6 +22,7 @@ Rails.application.config.to_prepare do
     Sidekiq::Cron::Job.create(name: 'Status Checking and Email Notification of Existing Batches - every 15min', cron: '*/15 * * * *', class: 'IngestBatchStatusEmailJobs::IngestFinished')
     Sidekiq::Cron::Job.create(name: 'Status Checking and Email Notification for Stalled Batches - every 1day', cron: '0 1 * * *', class: 'IngestBatchStatusEmailJobs::StalledJob')
     Sidekiq::Cron::Job.create(name: 'Clean out user sessions older than 7 days - every 6hour', cron: '0 */6 * * *', class: 'CleanupSessionJob')
+    Sidekiq::Cron::Job.create(name: 'Clean out expired stream tokens - every 5min', cron: '*/5 * * * *', class: 'CleanupStreamTokenJob')
   rescue Redis::CannotConnectError => e
     Rails.logger.warn "Cannot create sidekiq-cron jobs: #{e.message}"
   end
