@@ -1,6 +1,6 @@
 Rails.application.config.to_prepare do
-  HYDRANT_STEPS = Avalon::Workflow::Workflow.new(FileUploadStep.new,
-                                                 ResourceDescriptionStep.new,
-                                                 StructureStep.new,
-                                                 AccessControlStep.new)
+  HYDRANT_STEPS = Workflow.new(FileUploadStep.new,
+                               ResourceDescriptionStep.new,
+                               StructureStep.new,
+                               AccessControlStep.new)
 end
