@@ -297,7 +297,7 @@ module Avalon
         machine_generated = Avalon::Batch.true_field?(datastream[:machine_generated]) ? 'machine_generated' : nil
         # Create SupplementalFile
         supplemental_file = SupplementalFile.new(label: label, tags: [type, treat_as_transcript, machine_generated].uniq.compact, language: language, parent_id: parent_id)
-        supplemental_file.file.attach(io: FileLocator.new(datastream[file_key]).reader, filename: filename)
+        supplemental_file.attach_file(FileLocator.new(datastream[file_key]).reader, io: true)
         supplemental_file.save ? supplemental_file : nil
       end
       private_class_method :process_datastream
