@@ -26,29 +26,36 @@ FactoryBot.define do
       state { :running }
       percent_complete { 50.5 }
       current_operations { ['encoding'] }
-      input { FactoryBot.build(:encode_output) }
+      input { FactoryBot.build(:encode_input) }
     end
       
     trait :succeeded do
       state { :completed }
       percent_complete { 100 }
       current_operations { ['DONE'] }
-      input { FactoryBot.build(:encode_output) }
+      input { FactoryBot.build(:encode_input) }
       output { [ FactoryBot.build(:encode_output) ] }
+    end
+
+    trait :embedded_captions do
+      state { :completed }
+      percent_complete { 100 }
+      current_operations { ['DONE'] }
+      input { FactoryBot.build(:encode_input) }
+      output { [ FactoryBot.build(:encode_output), FactoryBot.build(:encode_supplemental_output) ] }
     end
 
     trait :failed do
       state { :failed }
       percent_complete { 50.5 }
       current_operations { ['FAILED'] }
-      input { FactoryBot.build(:encode_output) }
+      input { FactoryBot.build(:encode_input) }
       errors { ['Out of disk space.'] }
     end
   end
 
   factory :encode_input, class: ActiveEncode::Input do
     id { SecureRandom.uuid }
-    label { 'quality-high' }
     url { 'file://path/to/output.mp4' }
     duration { '21575.0' }
     audio_bitrate { '163842.0' }
@@ -57,6 +64,7 @@ FactoryBot.define do
     video_codec { 'AVC' }
     width { '1024' }
     height { '768' }
+    subtitles { [] }
   end
 
   factory :encode_output, class: ActiveEncode::Output do
@@ -70,5 +78,14 @@ FactoryBot.define do
     video_codec { 'AVC' }
     width { '1024' }
     height { '768' }
+    subtitles { [] }
+  end
+
+  factory :encode_supplemental_output, class: ActiveEncode::Output do
+    id { "gid://avalon/SupplementalFile/1" }
+    url { "file://#{Rails.root.join('spec', 'fixtures', 'caption.vtt')}"}
+    format { "vtt" }
+    label { "Test Caption" }
+    language { "en" }
   end
 end
