@@ -250,7 +250,7 @@ describe IiifPlaylistCanvasPresenter do
     subject { presenter.range }
 
     it 'generates the clip range' do
-    	expect(subject.label.to_s).to eq "{\"none\"=>[\"#{playlist_item.title}\"]}"
+    	expect(subject.label.to_s).to eq "{\"none\" => [\"#{playlist_item.title}\"]}"
     	expect(subject.items.size).to eq 1
     	expect(subject.items.first).to be_a IiifPlaylistCanvasPresenter
       expect(subject.items.first.media_fragment).to eq "t=#{playlist_item.start_time / 1000},#{playlist_item.end_time / 1000}"
